requirements.txt
Flask
uvicorn
azure-cosmos
python-dotenv